moved duplicated code from StepExecutionData in AbstractExecutionData
